Fix bottombar and pimp up the example
